refactor : error handling codes moved from daos to services

+++ services/commentService.js
@@ -6,6 +6,19 @@ const postComment = async (comment, id, user_id) => {
 };
 
 const modifiyComment = async (id, comment, user_id, comment_id) => {
+  const selectedComment = await commentDao.selectComment(comment_id);
+  //댓글이 존재하지 않을 경우 에러 발생
+  if (!selectedComment) {
+    const error = new Error('COMMENT DOES NOT EXIST');
+    error.statusCode = 404;
+    throw error;
+  }
+  //로그인한 사용자와 댓글 작성자가 다를 경우 에러 발생
+  if (selectedComment.user_id !== user_id) {
+    const error = new Error('ONLY WRITER CAN MODIFY COMMENT');
+    error.statusCode = 404;
+    throw error;
+  }
   const modifedComment = await commentDao.modifiyComment(
     id,
     comment,
@@ -16,6 +29,20 @@ const modifiyComment = async (id, comment, user_id, comment_id) => {
 };
 
 const deleteComment = async (user_id, comment_id) => {
+  const selectedComment = await commentDao.selectComment(comment_id);
+  //댓글이 존재하지 않을 경우 에러 발생
+  if (!selectedComment) {
+    const error = new Error('COMMENT DOES NOT EXIST');
+    error.statusCode = 404;
+    throw error;
+  }
+
+  //로그인한 사용자와 댓글 작성자가 다를 경우 에러 발생
+  if (selectedComment.user_id !== user_id) {
+    const error = new Error('ONLY WRITER CAN DELETE COMMENT');
+    error.statusCode = 404;
+    throw error;
+  }
   await commentDao.deleteComment(user_id, comment_id);
 };
